requirements.txt
pandas==1.1.1
pyyaml
spacy
pymongo
elasticsearch
vadersentiment
beautifulsoup4
pytesseract